requirements.txt
pandas
rembg
numpy>=1.25.0
requests
onnxruntime
pillow
pytesseract
deskew
realesrgan
basicsr
facexlib
gfpgan
torch
torchvision
torchaudio

# dev?
mypy
pytest
pytest-mypy
requests-mock